requirement.txt
attrs==19.1.0
aioamqp==0.12.0
